Added Illustrator enums and classes from ten-A/Types-for-Adobe
These are missing in the Illustrator typescript def docsforadobe/Types-for-adobe, but present in ten-A/Types-for-Adobe. I've taken a bunch of enums and classes from the latter, and added them to the former.

Note: there are still methods present in the ten-A definition that I haven't added. I just didn't want to go too crazy right off the bat.

## enum

- ComposerEngineType
- DiacVPosType
- DigitSetType
- DirOverrideType
- DocumentLayoutStyle
- DocumentType
- ElementPlacement
- ExportForScreensScaleType
- ExportForScreensType
- JPEGCompressionMethodType
- KashidasType
- KashidaWidthType
- ParagraphDirectionType
- PerspectiveGridType
- ViewRasterType
- ViewVectorType
- WariChuJustificationType

## class

- Asset
- Assets
- ExportForScreensItemToExport
- ExportForScreensOptionsJPEG
- ExportForScreensOptionsPNG24
- ExportForScreensOptionsPNG8
- ExportForScreensOptionsWebOptimizedSVG
- ExportForScreensPDFOptions
- MeshItem
- NonNativeItem